# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Interview Practice Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type StartInterviewBodyDomain =
  (typeof StartInterviewBodyDomain)[keyof typeof StartInterviewBodyDomain];

export const StartInterviewBodyDomain = {
  "software-developer": "software-developer",
  "data-science": "data-science",
  "ai-ml": "ai-ml",
  "hr-behavioral": "hr-behavioral",
  "general-aptitude": "general-aptitude",
} as const;

export interface StartInterviewBody {
  domain: StartInterviewBodyDomain;
  candidateName?: string;
}

export interface InterviewSession {
  sessionId: string;
  domain: string;
  firstQuestion: string;
  questionNumber: number;
  totalQuestions: number;
}

export interface TranscribeBody {
  /** Base64-encoded audio data */
  audio: string;
}

export interface TranscribeResponse {
  text: string;
}

export interface EvaluateBody {
  question: string;
  answer: string;
  domain: string;
  questionNumber: number;
}

export interface EvaluationResult {
  score: number;
  strengths: string;
  improvements: string;
  sampleBetterAnswer: string;
  feedbackText: string;
}

export interface NextQuestionBody {
  domain: string;
  questionNumber: number;
  previousQuestions: string[];
}

export interface QuestionResponse {
  question: string;
  questionNumber: number;
  isLast: boolean;
}

export type TTSBodyVoice = (typeof TTSBodyVoice)[keyof typeof TTSBodyVoice];

export const TTSBodyVoice = {
  alloy: "alloy",
  echo: "echo",
  fable: "fable",
  onyx: "onyx",
  nova: "nova",
  shimmer: "shimmer",
} as const;

export interface TTSBody {
  text: string;
  voice?: TTSBodyVoice;
}

export interface TTSResponse {
  /** Base64-encoded MP3 audio */
  audio: string;
}

export interface ReportBody {
  domain: string;
  evaluations: EvaluationResult[];
  questions: string[];
  /** Time taken in seconds */
  timeTaken?: number;
}

export interface PerformanceReport {
  overallScore: number;
  totalQuestions: number;
  grade: string;
  summary: string;
  strongAreas: string[];
  improvementAreas: string[];
  recommendations: string[];
}