# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Interview Practice Bot API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Start a new interview session
 */
export const StartInterviewBody = zod.object({
  domain: zod.enum([
    "software-developer",
    "data-science",
    "ai-ml",
    "hr-behavioral",
    "general-aptitude",
  ]),
  candidateName: zod.string().optional(),
});

export const StartInterviewResponse = zod.object({
  sessionId: zod.string(),
  domain: zod.string(),
  firstQuestion: zod.string(),
  questionNumber: zod.number(),
  totalQuestions: zod.number(),
});

/**
 * @summary Transcribe audio to text
 */
export const TranscribeAudioBody = zod.object({
  audio: zod.string().describe("Base64-encoded audio data"),
});

export const TranscribeAudioResponse = zod.object({
  text: zod.string(),
});

/**
 * @summary Evaluate an interview answer and get feedback
 */
export const EvaluateAnswerBody = zod.object({
  question: zod.string(),
  answer: zod.string(),
  domain: zod.string(),
  questionNumber: zod.number(),
});

export const EvaluateAnswerResponse = zod.object({
  score: zod.number(),
  strengths: zod.string(),
  improvements: zod.string(),
  sampleBetterAnswer: zod.string(),
  feedbackText: zod.string(),
});

/**
 * @summary Get the next interview question
 */
export const GetNextQuestionBody = zod.object({
  domain: zod.string(),
  questionNumber: zod.number(),
  previousQuestions: zod.array(zod.string()),
});

export const GetNextQuestionResponse = zod.object({
  question: zod.string(),
  questionNumber: zod.number(),
  isLast: zod.boolean(),
});

/**
 * @summary Convert text to speech audio
 */
export const TextToSpeechBody = zod.object({
  text: zod.string(),
  voice: zod
    .enum(["alloy", "echo", "fable", "onyx", "nova", "shimmer"])
    .optional(),
});

export const TextToSpeechResponse = zod.object({
  audio: zod.string().describe("Base64-encoded MP3 audio"),
});

/**
 * @summary Generate final performance report
 */
export const GenerateReportBody = zod.object({
  domain: zod.string(),
  evaluations: zod.array(
    zod.object({
      score: zod.number(),
      strengths: zod.string(),
      improvements: zod.string(),
      sampleBetterAnswer: zod.string(),
      feedbackText: zod.string(),
    }),
  ),
  questions: zod.array(zod.string()),
  timeTaken: zod.number().optional().describe("Time taken in seconds"),
});

export const GenerateReportResponse = zod.object({
  overallScore: zod.number(),
  totalQuestions: zod.number(),
  grade: zod.string(),
  summary: zod.string(),
  strongAreas: zod.array(zod.string()),
  improvementAreas: zod.array(zod.string()),
  recommendations: zod.array(zod.string()),
});