# lib/api-zod/src/generated/types/evaluateBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Interview Practice Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface EvaluateBody {
  question: string;
  answer: string;
  domain: string;
  questionNumber: number;
}