# lib/api-zod/src/generated/types/evaluationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Interview Practice Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface EvaluationResult {
  score: number;
  strengths: string;
  improvements: string;
  sampleBetterAnswer: string;
  feedbackText: string;
}