# lib/api-zod/src/generated/types/interviewSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Interview Practice Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface InterviewSession {
  sessionId: string;
  domain: string;
  firstQuestion: string;
  questionNumber: number;
  totalQuestions: number;
}