# lib/api-zod/src/generated/types/nextQuestionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Interview Practice Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface NextQuestionBody {
  domain: string;
  questionNumber: number;
  previousQuestions: string[];
}