# lib/api-zod/src/generated/types/performanceReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Interview Practice Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface PerformanceReport {
  overallScore: number;
  totalQuestions: number;
  grade: string;
  summary: string;
  strongAreas: string[];
  improvementAreas: string[];
  recommendations: string[];
}