# lib/api-zod/src/generated/types/reportBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Interview Practice Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { EvaluationResult } from "./evaluationResult";

export interface ReportBody {
  domain: string;
  evaluations: EvaluationResult[];
  questions: string[];
  /** Time taken in seconds */
  timeTaken?: number;
}