# lib/api-zod/src/generated/types/startInterviewBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Interview Practice Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { StartInterviewBodyDomain } from "./startInterviewBodyDomain";

export interface StartInterviewBody {
  domain: StartInterviewBodyDomain;
  candidateName?: string;
}